Move build.js and index.js to core as build-dev.js and build-static.js
