+++ app/react/package.json
@@ -80,9 +80,7 @@
     "url-loader": "^0.6.2",
     "util-deprecate": "^1.0.2",
     "uuid": "^3.2.1",
-    "webpack": "^3.10.0",
-    "webpack-dev-middleware": "^1.12.2",
-    "webpack-hot-middleware": "^2.21.0"
+    "webpack": "^3.10.0"
   },
   "devDependencies": {
     "nodemon": "^1.14.12"
