+++ app/react/src/server/utils.js
@@ -5,10 +5,6 @@ import deprecate from 'util-deprecate';
 const fallbackHeadUsage = deprecate(() => {},
 'Usage of head.html has been deprecated. Please rename head.html to preview-head.html');
 
-export function parseList(str) {
-  return str.split(',');
-}
-
 export function getPreviewHeadHtml(configDirPath) {
   const headHtmlPath = path.resolve(configDirPath, 'preview-head.html');
   const fallbackHtmlPath = path.resolve(configDirPath, 'head.html');
@@ -31,26 +27,4 @@ export function getManagerHeadHtml(configDirPath) {
   }
 
   return scriptHtml;
-}
-
-export function getEnvConfig(program, configEnv) {
-  Object.keys(configEnv).forEach(fieldName => {
-    const envVarName = configEnv[fieldName];
-    const envVarValue = process.env[envVarName];
-    if (envVarValue) {
-      program[fieldName] = envVarValue; // eslint-disable-line
-    }
-  });
-}
-
-export function getMiddleware(configDir) {
-  const middlewarePath = path.resolve(configDir, 'middleware.js');
-  if (fs.existsSync(middlewarePath)) {
-    let middlewareModule = require(middlewarePath); // eslint-disable-line
-    if (middlewareModule.__esModule) { // eslint-disable-line
-      middlewareModule = middlewareModule.default;
-    }
-    return middlewareModule;
-  }
-  return () => {};
 }
