+++ lib/core/package.json
@@ -20,12 +20,16 @@
     "@storybook/addons": "^3.4.0-alpha.7",
     "@storybook/channel-postmessage": "^3.4.0-alpha.7",
     "@storybook/client-logger": "^3.4.0-alpha.7",
+    "@storybook/node-logger": "^3.4.0-alpha.7",
     "@storybook/ui": "^3.4.0-alpha.7",
     "autoprefixer": "^7.2.5",
     "babel-runtime": "^6.26.0",
+    "chalk": "^2.3.0",
+    "commander": "^2.14.0",
     "css-loader": "^0.28.9",
     "dotenv": "^4.0.0",
     "events": "^1.1.1",
+    "express": "^4.16.2",
     "file-loader": "^1.1.6",
     "global": "^4.3.2",
     "json-loader": "^0.5.7",
@@ -35,8 +39,13 @@
     "qs": "^6.5.1",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
+    "serve-favicon": "^2.4.5",
+    "shelljs": "^0.8.1",
     "style-loader": "^0.20.1",
-    "url-loader": "^0.6.2"
+    "url-loader": "^0.6.2",
+    "webpack": "^3.10.0",
+    "webpack-dev-middleware": "^1.12.2",
+    "webpack-hot-middleware": "^2.21.0"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0"
