+++ lib/core/src/server/middleware.js
@@ -3,18 +3,17 @@ import { Router } from 'express';
 import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
-import getBaseConfig from './config/webpack.config';
-import loadConfig from './config';
 import { getMiddleware } from './utils';
 
 let webpackResolve = () => {};
 let webpackReject = () => {};
+
 export const webpackValid = new Promise((resolve, reject) => {
   webpackResolve = resolve;
   webpackReject = reject;
 });
 
-export default function(configDir) {
+export default function(configDir, loadConfig, getBaseConfig) {
   // Build the webpack configuration using the `getBaseConfig`
   // custom `.babelrc` file and `webpack.config.js` files
   const config = loadConfig('DEVELOPMENT', getBaseConfig(configDir), configDir);
